+++ pricing-service/src/main/java/com/destore/pricing/model/dto/PricingRuleResponse.java
@@ -28,13 +28,13 @@ import java.time.LocalDateTime;
 public class PricingRuleResponse {
 
     @Schema(description = "Unique rule identifier", example = "1")
-    private Integer ruleId;
+    private Long ruleId;
 
     @Schema(description = "Item ID from warehouse", example = "1")
     private Integer itemId;
 
     @Schema(description = "Store ID (null = global rule)", example = "2")
-    private Integer storeId;
+    private Long storeId;
 
     @Schema(description = "Store code", example = "LON-001")
     private String storeCode;
